[feat] points 클래스 추가
